Move ViewModelAuthSupport to session
