+++ app/shared/src/commonMain/kotlin/session/isLoggedIn.kt
@@ -5,7 +5,6 @@ import androidx.compose.runtime.Stable
 import androidx.compose.runtime.State
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.filterNotNull
-import me.him188.ani.app.ViewModelAuthSupport
 import me.him188.ani.app.ui.foundation.AbstractViewModel
 import me.him188.ani.app.ui.foundation.rememberViewModel
 import moe.tlaster.precompose.flow.collectAsStateWithLifecycle
