+++ app/shared/src/commonMain/kotlin/ui/main/HomeScene.kt
@@ -50,8 +50,8 @@ import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
 import me.him188.ani.app.navigation.LocalNavigator
 import me.him188.ani.app.platform.LocalContext
-import me.him188.ani.app.requireOnline
 import me.him188.ani.app.session.isLoggedIn
+import me.him188.ani.app.session.requireOnline
 import me.him188.ani.app.tools.update.InstallationFailureReason
 import me.him188.ani.app.ui.cache.CacheManagementPage
 import me.him188.ani.app.ui.external.placeholder.placeholder
