+++ app/shared/src/commonMain/kotlin/ui/profile/AccountViewModel.kt
@@ -5,9 +5,9 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import kotlinx.coroutines.flow.map
-import me.him188.ani.app.ViewModelAuthSupport
 import me.him188.ani.app.data.repositories.ProfileRepository
 import me.him188.ani.app.session.SessionManager
+import me.him188.ani.app.session.ViewModelAuthSupport
 import me.him188.ani.app.ui.foundation.AbstractViewModel
 import me.him188.ani.app.ui.foundation.launchInBackground
 import me.him188.ani.utils.coroutines.runUntilSuccess
